+++ src/Symfony/Bridge/Twig/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 3.2.0
 -----
 
+ * added `AppVariable::getToken()`
  * Deprecated the possibility to inject the Form Twig Renderer into the form
    extension. Inject it on TwigRendererEngine instead.